+++ changelog.md
@@ -3,6 +3,7 @@
 - Removed **Edit Camera** due to issues with crashes on levels and compliances with the Geode Index guidelines. sorry :(
 - Fixed TextBox inputs not allowing user to set an decimal point on certain actions
 - Added **Set Gravity**,**Speed Player** & **Noclip** Events
+- Added a delete command comfirmation popup
 # v0.1.5-beta.1
 - Added **Sound Effect** & **Stop All Sounds** Events
 - Changed the **Wait** Action value to use float value instead of just integer.